+++ source/LibraryCleaner/library_cleaner/__init__.py
@@ -1,3 +1,3 @@
 """RadioBOSS Library Cleaner."""
 
-__version__ = "1.0.0"
+__version__ = "1.0.1"